refactor: 💡 move to pnpm
