+++ .eslintignore
@@ -6,18 +6,6 @@ expected
 website
 gh-pages
 vendors
-packages/build-scripts/test
-packages/template-plugin/tests
-packages/build-scripts/bin
-packages/build-scripts/lib
-packages/build-plugin-store/lib
-packages/build-plugin-ice-router/lib
-packages/build-plugin-ice-stark-child/lib
-packages/build-plugin-env-config/lib
-packages/build-plugin-webpack5/lib
-packages/build-plugin-rml/lib
-packages/build-plugin-stark-module/lib
-packages/build-plugin-stark-module/test
-packages/template-component-demo/lib
-packages/template-component-demo/es
 examples
+__tests__
+lib
