+++ package.json
@@ -4,8 +4,9 @@
     "packages/*"
   ],
   "scripts": {
+    "preinstall": "npx only-allow pnpm",
     "clean": "rm -rf node_modules ./packages/*/node_modules",
-    "setup": "npm run clean && yarn install --registry=https://registry.npm.taobao.org && lerna bootstrap --registry=https://registry.npm.taobao.org",
+    "setup": "npm run clean && pnpm i --registry=https://registry.npmmirror.com",
     "lint": "eslint --cache --quiet --ext .js,.ts ./",
     "lint-fix": "eslint --cache --ext .js,.ts ./ --fix",
     "prettier": "prettier --write '**/*.{js,jsx,tsx,ts,less,md,json}'",
@@ -22,9 +23,9 @@
     "@commitlint/cli": "^7.5.2",
     "@commitlint/config-conventional": "^7.5.0",
     "@iceworks/spec": "^1.5.0",
+    "build-scripts": "workspace:*",
     "eslint": "^7.31.0",
     "husky": "^1.3.1",
-    "lerna": "^3.16.4",
     "stylelint": "^10.1.0",
     "typescript": "^4"
   }
