+++ packages/build-scripts/__tests__/configFile/config.js
