+++ packages/build-scripts/__tests__/configFile/esmConfig.js
