+++ packages/build-scripts/__tests__/configFiles.test.ts
