+++ packages/build-scripts/__tests__/fixtures/basic/build.json
