+++ packages/build-scripts/__tests__/fixtures/jsConfig/build.config.js
