+++ packages/build-scripts/__tests__/fixtures/mixConfig/build.config.js
