+++ packages/build-scripts/__tests__/fixtures/mixConfig/build.json
