+++ packages/build-scripts/__tests__/fixtures/modeConfig/a.plugin.js
