+++ packages/build-scripts/__tests__/fixtures/modeConfig/b.plugin.js
