+++ packages/build-scripts/__tests__/fixtures/modeConfig/build.json
