+++ packages/build-scripts/__tests__/fixtures/tsConfig/build.config.ts
