+++ packages/build-scripts/__tests__/fixtures/userConfig/build.json
