+++ packages/build-scripts/__tests__/loadConfig.test.ts
