+++ packages/build-scripts/src/index.ts
@@ -1,9 +1,12 @@
 import Service from './Service';
 import Context, { createContext } from './Context';
+
+export type { ITaskConfig, IJestResult } from './Context';
+
 export * from './types';
 
 export {
   Service,
   Context,
   createContext,
-}
+};
