+++ packages/build-scripts/src/types/index.ts
@@ -1,3 +1,9 @@
+import { AggregatedResult } from '@jest/test-result';
+import { GlobalConfig } from '@jest/types/build/Config';
+import { PLUGIN_CONTEXT_KEY, VALIDATION_MAP } from '../utils/constant';
+import type { CreateLoggerReturns } from '../utils/logger';
+import type { Context } from '..';
+
 export interface IHash<T> {
   [name: string]: T;
 }
@@ -11,7 +17,3 @@ export type JsonValue = Json[keyof Json];
 export type MaybeArray<T> = T | T[];
 
 export type MaybePromise<T> = T | Promise<T>;
-
-export interface IRunOptions {
-  eject?: boolean;
-}
