+++ packages/template-plugin/__tests__/build.spec.ts
