+++ packages/template-plugin/__tests__/fixtures/basic-spa/build.json
