+++ packages/template-plugin/__tests__/fixtures/basic-spa/src/index.tsx
