+++ packages/template-plugin/__tests__/fixtures/defaultConfig.ts
