+++ packages/template-plugin/__tests__/index.spec.ts
