+++ packages/webpack-service/src/types.ts
@@ -1,0 +1,3 @@
+export interface IRunOptions {
+  eject?: boolean;
+}
