+++ packages/webpack-service/src/utils/log.ts
@@ -1,0 +1,13 @@
+import npmlog from 'npmlog';
+
+const envs = ['verbose', 'info', 'error', 'warn'];
+const logLevel =
+  envs.indexOf(process.env.LOG_LEVEL) !== -1 ? process.env.LOG_LEVEL : 'info';
+
+npmlog.level = logLevel;
+
+// LOG_LEVEL=verbose for debug
+// log.verbose
+// log.info
+// log.error
+export default npmlog;
